@@ -16,7 +16,7 @@
         }
     ],
     "require": {
-        "php": "^7.1",
+        "php": "^7.4",
         "container-interop/container-interop": "^1",
         "psr/container": "^1.0"
     },
