@@ -9,24 +9,23 @@ use Midnight\Permissions\Exception\InvalidPermissionException;
 use Midnight\Permissions\Exception\UnknownPermissionException;
 use Midnight\Permissions\PermissionInterface;
 use Midnight\Permissions\PermissionService;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use PHPUnit_Framework_MockObject_MockObject;
 use stdClass;
 
 class PermissionServiceTest extends TestCase
 {
-    /** @var ContainerInterface|PHPUnit_Framework_MockObject_MockObject */
+    /** @var ContainerInterface & MockObject */
     private $container;
-    /** @var PermissionService */
-    private $service;
+    private PermissionService $service;
 
-    public function setUp()
+    protected function setUp(): void
     {
         $this->container = $this->createMock(ContainerInterface::class);
         $this->service = new PermissionService($this->container);
     }
 
-    public function testIsAllowed()
+    public function testIsAllowed(): void
     {
         $user = new stdClass();
         $permissionName = 'can_do_something';
@@ -34,7 +33,7 @@ class PermissionServiceTest extends TestCase
 
         $permission = $this->createMock(PermissionInterface::class);
         $permission
-            ->expects($this->any())
+            ->expects(self::any())
             ->method('isAllowed')
             ->with($user, $resource)
             ->willReturn(true);
@@ -42,10 +41,10 @@ class PermissionServiceTest extends TestCase
         $this->container->method('has')->willReturn(true);
         $this->container->method('get')->willReturn($permission);
 
-        $this->assertTrue($this->service->isAllowed($user, $permissionName, $resource));
+        self::assertTrue($this->service->isAllowed($user, $permissionName, $resource));
     }
 
-    public function testUnknownPermission()
+    public function testUnknownPermission(): void
     {
         $this->container->method('has')->willReturn(false);
 
@@ -54,7 +53,7 @@ class PermissionServiceTest extends TestCase
         $this->service->isAllowed(new stdClass(), 'some_unknown_permission', new stdClass());
     }
 
-    public function testInvalidPermission()
+    public function testInvalidPermission(): void
     {
         $this->container->method('has')->willReturn(true);
         $this->container->method('get')->willReturn(new stdClass());
